+++ enterprise/internal/insights/gitserver/first_commit.go
@@ -1,4 +1,4 @@
-package discovery
+package gitserver
 
 import (
 	"context"
