+++ enterprise/internal/insights/gitserver/first_commit_test.go
@@ -1,4 +1,4 @@
-package discovery
+package gitserver
 
 import (
 	"testing"
